RAD-233 Move MwlStatus to package study
