+++ api/src/main/resources/RadiologyStudy.hbm.xml
@@ -35,7 +35,7 @@
 		<property name="mwlStatus" column="mwl_status" length="17"
 			not-null="true">
 			<type name="org.openmrs.util.HibernateEnumType">
-				<param name="enumClassName">org.openmrs.module.radiology.MwlStatus</param>
+				<param name="enumClassName">org.openmrs.module.radiology.study.MwlStatus</param>
 			</type>
 		</property>
 	</class>
