+++ api/src/test/java/org/openmrs/module/radiology/order/RadiologyOrderServiceImplComponentTest.java
@@ -37,9 +37,9 @@ import org.openmrs.api.PatientService;
 import org.openmrs.api.ProviderService;
 import org.openmrs.api.VisitService;
 import org.openmrs.module.emrapi.encounter.EmrEncounterService;
-import org.openmrs.module.radiology.MwlStatus;
 import org.openmrs.module.radiology.RadiologyProperties;
 import org.openmrs.module.radiology.study.RadiologyStudyService;
+import org.openmrs.module.radiology.study.MwlStatus;
 import org.openmrs.module.radiology.study.RadiologyStudy;
 import org.openmrs.test.BaseModuleContextSensitiveTest;
 import org.springframework.beans.factory.annotation.Autowired;
